requirements.txt
python-telegram-bot==12.3.0
spotdl
python-dotenv
urllib3
